Added account image delete logic

+++ app/Actions/User/DeleteUserAction.php
@@ -2,7 +2,9 @@
 
 namespace App\Actions\User;
 
+use App\Enums\UserRole;
 use App\Models\User;
+use Illuminate\Validation\ValidationException;
 
 class DeleteUserAction
 {
@@ -14,6 +16,12 @@ class DeleteUserAction
      */
     public function handle(User $user): void
     {
+        if ($user->hasRole(UserRole::SUPER_ADMIN->value)) {
+            throw ValidationException::withMessages([
+                'email' => [__('users.protected')],
+            ]);
+        }
+
         $user->delete();
     }
 }
